requirements.txt
Flask==2.2.3
flask-socketio==5.3.2
flask_cors==3.0.10
Pillow==9.4.0